Fixes to deal with failed tests
Yes, the initial round of testing went poorly.

--- PYTEST PYPEIT UNIT TESTS PASSED  261 passed, 445 warnings in 541.44s (0:09:01) ---
--- PYTEST UNIT TESTS FAILED  3 failed, 154 passed, 126104 warnings in 1040.84s (0:17:20) ---
--- PYTEST VET TESTS FAILED  20 failed, 47 passed, 142 warnings in 4640.21s (1:17:20) ---
--- PYPEIT DEVELOPMENT SUITE FAILED 52/255 TESTS  ---

This commit addresses most of the errors found when going through the
report, but there will still be some errors.  In the next cloud run,
I will be copying out all the "*.log" files to be able to see more than
the last line of the log for what happened.

Of these fixes, most have to do with making sure metadata keys needed
for ``config_specific_par()`` are actually in the PypeIt Reduction File
or are configuration keys.

The NOT/NTT errors were related to "#" characters in the grating name
that get scrubbed by the metadata class.  And a couple of things I just
plain missed the first time through.

Here's hoping!  🤞

	modified:   pypeit/scripts/chk_for_calibs.py
	modified:   pypeit/spectrographs/gemini_gnirs.py
	modified:   pypeit/spectrographs/gtc_osiris.py
	modified:   pypeit/spectrographs/keck_deimos.py
	modified:   pypeit/spectrographs/keck_lris.py
	modified:   pypeit/spectrographs/keck_mosfire.py
	modified:   pypeit/spectrographs/ldt_deveny.py
	modified:   pypeit/spectrographs/mmt_bluechannel.py
	modified:   pypeit/spectrographs/not_alfosc.py
	modified:   pypeit/spectrographs/ntt_efosc2.py
	modified:   pypeit/spectrographs/shane_kast.py