Operational and typographic fixes
Fold the recent APF/Levy binning update into this scheme, fix the
Keck/MOSFIRE duplicate column and missed typos.

These should be the last changes here needed to make all tests pass.
What's next is to propagate the new ``configuration_keys()`` columns
into the ``.pypeit`` files in the DevSuite that are used for the
instrument tests.  This mismatch between older PypeIt Reduction Files
and what is expected here caused most of the failues in the last run.

--- PYTEST PYPEIT UNIT TESTS PASSED  261 passed, 441 warnings in 414.78s (0:06:54) ---
--- PYTEST UNIT TESTS FAILED  5 failed, 152 passed, 95145 warnings in 952.64s (0:15:52) ---
--- PYTEST VET TESTS FAILED  17 failed, 50 passed, 490 warnings in 4439.25s (1:13:59) ---
--- PYPEIT DEVELOPMENT SUITE FAILED 31/259 TESTS  ---

	modified:   pypeit/spectrographs/apf_levy.py
	modified:   pypeit/spectrographs/keck_deimos.py
	modified:   pypeit/spectrographs/keck_lris.py
	modified:   pypeit/spectrographs/keck_mosfire.py
	modified:   pypeit/spectrographs/shane_kast.py