Add new FindObj par ``trace_min_max``
This parameter feeds the extant argument in ``findobj_skymask.objs_in_slit()``
called ``spec_min_max`` that limits the spectral range (measured in pixels)
over which the trace is fix.  This is useful for situations where the spectral
range on the detector exceeds the region where sky photons can be detected
(either atmospheric or silicon cutoffs or both).  The default value is ``None``
which corresponds to previous behavior of using the entire chip.

Conformal changes.

	modified:   pypeit/coadd2d.py
	modified:   pypeit/core/findobj_skymask.py
	modified:   pypeit/find_objects.py
	modified:   pypeit/par/pypeitpar.py
	modified:   pypeit/spectrographs/ldt_deveny.py